Move built in validation into a separate file
